f/WPC-732 add metadata detection on pload in article and page
