+++ packages/editor/src/client/panel/imageSelectPanel.tsx
@@ -21,6 +21,7 @@ import {Typography} from '../atoms/typography'
 import {getImgMinSizeToCompress} from '../utility'
 import {ImagedEditPanel} from './imageEditPanel'
 import {createCheckedPermissionComponent} from '../atoms/permissionControl'
+import {ImageMetaData, readImageMetaData} from '../atoms/imageMetaData'
 
 export interface ImageSelectPanelProps {
   onClose(): void
@@ -33,6 +34,13 @@ function ImageSelectPanel({onClose, onSelect}: ImageSelectPanelProps) {
   const [filter, setFilter] = useState('')
 
   const [file, setFile] = useState<File | null>(null)
+  const [imageMetaData, setImageMetaData] = useState<ImageMetaData>({
+    title: '',
+    description: '',
+    source: '',
+    link: '',
+    licence: ''
+  })
   const {data, fetchMore, loading: isLoading} = useImageListQuery({
     fetchPolicy: 'network-only',
     variables: {
@@ -49,6 +57,8 @@ function ImageSelectPanel({onClose, onSelect}: ImageSelectPanelProps) {
     if (files.length === 0) return
 
     const file = files[0]
+
+    setImageMetaData(await readImageMetaData(file))
 
     if (!file.type.startsWith('image')) {
       toaster.push(
@@ -88,6 +98,7 @@ function ImageSelectPanel({onClose, onSelect}: ImageSelectPanelProps) {
         onClose={onClose}
         file={file}
         onSave={(image: ImageRefFragment) => onSelect(image)}
+        imageMetaData={imageMetaData}
       />
     )
   }
